PYIC-3506 Added few tests, refactor package.
